update of script

+++ scripts/vm_scripts/vm_preparation_to_build_dynamic_and_core_agent_from_github.sh
@@ -4,6 +4,11 @@
 GIT_DIRECTORY=/home/nao/download/git
 WS_RAPP_DIRECTORY=/home/nao/ws_rapp/src
 
+if [ "$#" -ne 1 ]; then
+  echo "Usage: $0 <flag>"
+  exit 1
+fi
+
 if [ -d $GIT_DIRECTORY ]; then #If direcotory exists
   echo "Removing - a directory. Cleaning git folder."
   rm $GIT_DIRECTORY -rf
@@ -18,9 +23,14 @@ else
   ~/my_workspace/src/catkin/bin/catkin_init_workspace
 fi
 
-mkdir -p ~/download/git
-cd ~/download/git
-git clone https://github.com/rapp-project/rapp-robot-nao.git
+if [ $1 -eq 0 ]; then #clone from github
+  echo "Clonning from github"
+  mkdir -p ~/download/git
+  cd ~/download/git
+  git clone https://github.com/rapp-project/rapp-robot-nao.git
+fi
+
+
 cp -r ~/download/git/rapp-robot-nao/scripts/vm_scripts/* ~/scripts
 cp -r ~/download/git/rapp-robot-nao/rapp_robot_agent ~/ws_rapp/src
 cp -r ~/download/git/rapp-robot-nao/rapp_libraries ~/ws_rapp/src #change to RAPP API
